pyproject.toml
[project]
name = "pydantic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai-slim[openai]",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation",
    "opentelemetry-instrumentation-requests",
    "opentelemetry-instrumentation-httpx",
    "langsmith",
    "python-dotenv",
]
